Extract `CardanoFrag` to types.ts
